+++ relayer/chains/icon/provider.go
@@ -6,28 +6,19 @@ import (
 	"math/big"
 
 	"github.com/icon-project/centralized-relay/relayer/chains/icon/types"
+	"github.com/icon-project/centralized-relay/relayer/events"
 	"github.com/icon-project/centralized-relay/relayer/kms"
 	"github.com/icon-project/centralized-relay/relayer/provider"
 	providerTypes "github.com/icon-project/centralized-relay/relayer/types"
-	relayerTypes "github.com/icon-project/centralized-relay/relayer/types"
 	"github.com/icon-project/goloop/module"
 	"go.uber.org/zap"
 )
 
 type Config struct {
-	ChainName      string                         `json:"-" yaml:"-"`
-	RPCUrl         string                         `json:"rpc-url" yaml:"rpc-url"`
-	Address        string                         `json:"address" yaml:"address"`
-	StartHeight    uint64                         `json:"start-height" yaml:"start-height"` // would be of highest priority
-	Contracts      relayerTypes.ContractConfigMap `json:"contracts" yaml:"contracts"`
-	NetworkID      int64                          `json:"network-id" yaml:"network-id"`
-	FinalityBlock  uint64                         `json:"finality-block" yaml:"finality-block"`
-	NID            string                         `json:"nid" yaml:"nid"`
-	StepMin        int64                          `json:"step-min" yaml:"step-min"`
-	StepLimit      int64                          `json:"step-limit" yaml:"step-limit"`
-	StepAdjustment int64                          `json:"step-adjustment" yaml:"step-adjustment"`
-	Disabled       bool                           `json:"disabled" yaml:"disabled"`
-	HomeDir        string                         `json:"-" yaml:"-"`
+	provider.CommonConfig `json:",inline" yaml:",inline"`
+	StepMin               int64 `json:"step-min" yaml:"step-min"`
+	StepLimit             int64 `json:"step-limit" yaml:"step-limit"`
+	StepAdjustment        int64 `json:"step-adjustment" yaml:"step-adjustment"`
 }
 
 // NewProvider returns new Icon provider
@@ -39,13 +30,20 @@ func (c *Config) NewProvider(ctx context.Context, log *zap.Logger, homepath stri
 		return nil, err
 	}
 
+	client := NewClient(ctx, c.RPCUrl, log)
+	NetworkInfo, err := client.GetNetworkInfo()
+	if err != nil {
+		return nil, fmt.Errorf("failed to get network id: %v", err)
+	}
+
 	c.ChainName = chainName
 	c.HomeDir = homepath
 
 	return &Provider{
 		log:       log.With(zap.Stringp("nid ", &c.NID), zap.Stringp("name", &c.ChainName)),
-		client:    NewClient(ctx, c.RPCUrl, log),
+		client:    client,
 		cfg:       c,
+		networkID: NetworkInfo.NetworkID,
 		contracts: c.eventMap(),
 	}, nil
 }
@@ -93,6 +91,7 @@ type Provider struct {
 	client              *Client
 	kms                 kms.KMS
 	contracts           map[string]providerTypes.EventMap
+	networkID           types.HexInt
 	LastSavedHeightFunc func() uint64
 }
 
@@ -117,6 +116,10 @@ func (p *Provider) Name() string {
 	return p.cfg.ChainName
 }
 
+func (p *Provider) NetworkID() types.HexInt {
+	return p.networkID
+}
+
 func (p *Provider) Wallet() (module.Wallet, error) {
 	if p.wallet == nil {
 		if err := p.RestoreKeystore(context.Background()); err != nil {
@@ -132,21 +135,26 @@ func (p *Provider) FinalityBlock(ctx context.Context) uint64 {
 
 // MessageReceived checks if the message is received
 func (p *Provider) MessageReceived(ctx context.Context, messageKey *providerTypes.MessageKey) (bool, error) {
-	callParam := p.prepareCallParams(MethodGetReceipts, p.cfg.Contracts[providerTypes.ConnectionContract], map[string]interface{}{
-		"srcNetwork": messageKey.Src,
-		"_connSn":    types.NewHexInt(int64(messageKey.Sn)),
-	})
-
-	var status types.HexInt
-	if err := p.client.Call(callParam, &status); err != nil {
-		return false, fmt.Errorf("MessageReceived: %v", err)
+	switch messageKey.EventType {
+	case events.EmitMessage:
+		callParam := p.prepareCallParams(MethodGetReceipts, p.cfg.Contracts[providerTypes.ConnectionContract], map[string]interface{}{
+			"srcNetwork": messageKey.Src,
+			"_connSn":    types.NewHexInt(messageKey.Sn.Int64()),
+		})
+		var status types.HexInt
+		return status == types.NewHexInt(1), p.client.Call(callParam, &status)
+	case events.CallMessage:
+		return false, nil
+	case events.RollbackMessage:
+		return false, nil
+	default:
+		return true, fmt.Errorf("unknown event type")
 	}
-	return status == types.NewHexInt(1), nil
 }
 
 // ReverseMessage reverts a message
 func (p *Provider) RevertMessage(ctx context.Context, sn *big.Int) error {
-	params := map[string]interface{}{"_sn": types.NewHexInt(sn.Int64())}
+	params := map[string]interface{}{"sn": types.NewHexInt(sn.Int64())}
 	message := p.NewIconMessage(types.Address(p.cfg.Contracts[providerTypes.ConnectionContract]), params, MethodRevertMessage)
 	txHash, err := p.SendTransaction(ctx, message)
 	if err != nil {
@@ -202,11 +210,11 @@ func (p *Provider) GetFee(ctx context.Context, networkID string, responseFee boo
 }
 
 // SetFees
-func (p *Provider) SetFee(ctx context.Context, networkID string, msgFee, resFee uint64) error {
+func (p *Provider) SetFee(ctx context.Context, networkID string, msgFee, resFee *big.Int) error {
 	callParam := map[string]interface{}{
 		"networkId":   networkID,
-		"messageFee":  types.NewHexInt(int64(msgFee)),
-		"responseFee": types.NewHexInt(int64(resFee)),
+		"messageFee":  types.NewHexInt(msgFee.Int64()),
+		"responseFee": types.NewHexInt(resFee.Int64()),
 	}
 
 	msg := p.NewIconMessage(types.Address(p.cfg.Contracts[providerTypes.ConnectionContract]), callParam, MethodSetFee)
@@ -243,8 +251,8 @@ func (p *Provider) ClaimFee(ctx context.Context) error {
 }
 
 // ExecuteRollback
-func (p *Provider) ExecuteRollback(ctx context.Context, sn uint64) error {
-	params := map[string]interface{}{"_sn": types.NewHexInt(int64(sn))}
+func (p *Provider) ExecuteRollback(ctx context.Context, sn int64) error {
+	params := map[string]interface{}{"_sn": types.NewHexInt(sn)}
 	message := p.NewIconMessage(types.Address(p.cfg.Contracts[providerTypes.XcallContract]), params, MethodExecuteRollback)
 	txHash, err := p.SendTransaction(ctx, message)
 	if err != nil {