+++ relayer/chains/wasm/provider.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"math/big"
 	"runtime"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -34,6 +33,7 @@ type Provider struct {
 	contracts           map[string]relayTypes.EventMap
 	eventList           []sdkTypes.Event
 	LastSavedHeightFunc func() uint64
+	routerMutex         *sync.Mutex
 }
 
 func (p *Provider) QueryLatestHeight(ctx context.Context) (uint64, error) {
@@ -70,10 +70,10 @@ func (p *Provider) Init(ctx context.Context, homePath string, kms kms.KMS) error
 
 // Wallet returns the wallet of the provider
 func (p *Provider) Wallet() sdkTypes.AccAddress {
-	if p.wallet == nil {
-		ctx := context.Background()
-		done := p.SetSDKContext()
-		defer done()
+	ctx := context.Background()
+	done := p.SetSDKContext()
+	defer done()
+	if p.wallet == nil || p.wallet.GetAddress().Empty() {
 		if err := p.RestoreKeystore(ctx); err != nil {
 			p.logger.Error("failed to restore keystore", zap.Error(err))
 			return nil
@@ -128,11 +128,13 @@ func (p *Provider) Listener(ctx context.Context, lastSavedHeight uint64, blockIn
 		case <-subscribeStarter.C:
 			subscribeStarter.Stop()
 			for _, event := range p.contracts {
-				go p.SubscribeMessageEvents(ctx, blockInfoChan, &types.SubscribeOpts{
-					Address: event.Address,
-					Method:  event.GetWasmMsgType(),
-					Height:  latestHeight,
-				}, resetFunc)
+				for msgType := range event.SigType {
+					go p.SubscribeMessageEvents(ctx, blockInfoChan, &types.SubscribeOpts{
+						Address: event.Address,
+						Method:  msgType,
+						Height:  latestHeight,
+					}, resetFunc)
+				}
 			}
 		case <-pollHeightTicker.C:
 			pollHeightTicker.Stop()
@@ -140,7 +142,7 @@ func (p *Provider) Listener(ctx context.Context, lastSavedHeight uint64, blockIn
 			latestHeight, err = p.QueryLatestHeight(ctx)
 			if err != nil {
 				p.logger.Error("failed to get latest block height", zap.Error(err))
-				continue
+				pollHeightTicker.Reset(time.Second * 3)
 			}
 		default:
 			if startHeight < latestHeight {
@@ -161,8 +163,7 @@ func (p *Provider) Route(ctx context.Context, message *relayTypes.Message, callb
 	if err := p.wallet.SetSequence(seq); err != nil {
 		p.logger.Error("failed to set sequence", zap.Error(err))
 	}
-	p.waitForTxResult(ctx, message.MessageKey(), res.TxHash, callback)
-	return nil
+	return p.waitForTxResult(ctx, message.MessageKey(), res, callback)
 }
 
 // call the smart contract to send the message
@@ -177,7 +178,7 @@ func (p *Provider) call(ctx context.Context, message *relayTypes.Message) (*sdkT
 	switch message.EventType {
 	case events.EmitMessage, events.RevertMessage, events.SetAdmin, events.ClaimFee, events.SetFee:
 		contract = p.cfg.Contracts[relayTypes.ConnectionContract]
-	case events.CallMessage, events.ExecuteRollback:
+	case events.CallMessage, events.RollbackMessage:
 		contract = p.cfg.Contracts[relayTypes.XcallContract]
 	default:
 		return nil, fmt.Errorf("unknown event type: %s ", message.EventType)
@@ -204,6 +205,8 @@ func (p *Provider) call(ctx context.Context, message *relayTypes.Message) (*sdkT
 }
 
 func (p *Provider) sendMessage(ctx context.Context, msgs ...sdkTypes.Msg) (*sdkTypes.TxResponse, error) {
+	p.routerMutex.Lock()
+	defer p.routerMutex.Unlock()
 	return p.prepareAndPushTxToMemPool(ctx, p.wallet.GetAccountNumber(), p.wallet.GetSequence(), msgs...)
 }
 
@@ -215,19 +218,18 @@ func (p *Provider) handleSequence(ctx context.Context) error {
 	return p.wallet.SetSequence(acc.GetSequence())
 }
 
-func (p *Provider) logTxFailed(err error, txHash string, code uint8) {
+func (p *Provider) logTxFailed(err error, tx *sdkTypes.TxResponse) {
 	p.logger.Error("transaction failed",
+		zap.String("tx_hash", tx.TxHash),
+		zap.String("codespace", tx.Codespace),
 		zap.Error(err),
-		zap.String("tx_hash", txHash),
-		zap.Uint8("code", code),
 	)
 }
 
-func (p *Provider) logTxSuccess(height uint64, txHash string) {
+func (p *Provider) logTxSuccess(res *types.TxResult) {
 	p.logger.Info("successful transaction",
-		zap.Uint64("block_height", height),
-		zap.String("chain_id", p.cfg.ChainID),
-		zap.String("tx_hash", txHash),
+		zap.Int64("block_height", res.TxResult.Height),
+		zap.String("tx_hash", res.TxResult.TxHash),
 	)
 }
 
@@ -275,27 +277,27 @@ func (p *Provider) prepareAndPushTxToMemPool(ctx context.Context, acc, seq uint6
 	return res, nil
 }
 
-func (p *Provider) waitForTxResult(ctx context.Context, mk *relayTypes.MessageKey, txHash string, callback relayTypes.TxResponseFunc) {
-	for txWaitRes := range p.subscribeTxResultStream(ctx, txHash, p.cfg.TxConfirmationInterval) {
-		if txWaitRes.Error != nil && txWaitRes.Error != context.DeadlineExceeded {
-			p.logTxFailed(txWaitRes.Error, txHash, uint8(txWaitRes.TxResult.Code))
-			callback(mk, txWaitRes.TxResult, txWaitRes.Error)
-			return
-		}
-		p.logTxSuccess(uint64(txWaitRes.TxResult.Height), txHash)
-		callback(mk, txWaitRes.TxResult, nil)
+func (p *Provider) waitForTxResult(ctx context.Context, mk *relayTypes.MessageKey, tx *sdkTypes.TxResponse, callback relayTypes.TxResponseFunc) error {
+	res, err := p.subscribeTxResult(ctx, tx, p.cfg.TxConfirmationInterval)
+	if err != nil {
+		callback(mk, res.TxResult, err)
+		p.logTxFailed(err, tx)
+		return err
 	}
+	callback(mk, res.TxResult, nil)
+	p.logTxSuccess(res)
+	return nil
 }
 
-func (p *Provider) pollTxResultStream(ctx context.Context, txHash string, maxWaitInterval time.Duration) <-chan *types.TxResultChan {
-	txResChan := make(chan *types.TxResultChan)
+func (p *Provider) pollTxResultStream(ctx context.Context, txHash string, maxWaitInterval time.Duration) <-chan *types.TxResult {
+	txResChan := make(chan *types.TxResult)
 	startTime := time.Now()
-	go func(txChan chan *types.TxResultChan) {
+	go func(txChan chan *types.TxResult) {
 		defer close(txChan)
 		for range time.NewTicker(p.cfg.TxConfirmationInterval).C {
 			res, err := p.client.GetTransactionReceipt(ctx, txHash)
 			if err == nil {
-				txChan <- &types.TxResultChan{
+				txChan <- &types.TxResult{
 					TxResult: &relayTypes.TxResponse{
 						Height:    res.TxResponse.Height,
 						TxHash:    res.TxResponse.TxHash,
@@ -306,7 +308,7 @@ func (p *Provider) pollTxResultStream(ctx context.Context, txHash string, maxWai
 				}
 				return
 			} else if time.Since(startTime) > maxWaitInterval {
-				txChan <- &types.TxResultChan{
+				txChan <- &types.TxResult{
 					Error: err,
 				}
 				return
@@ -316,101 +318,83 @@ func (p *Provider) pollTxResultStream(ctx context.Context, txHash string, maxWai
 	return txResChan
 }
 
-func (p *Provider) subscribeTxResultStream(ctx context.Context, txHash string, maxWaitInterval time.Duration) <-chan *types.TxResultChan {
-	txResChan := make(chan *types.TxResultChan)
-	go func(txRes chan *types.TxResultChan) {
-		defer close(txRes)
+func (p *Provider) subscribeTxResult(ctx context.Context, tx *sdkTypes.TxResponse, maxWaitInterval time.Duration) (*types.TxResult, error) {
+	newCtx, cancel := context.WithTimeout(ctx, maxWaitInterval)
+	defer cancel()
 
-		newCtx, cancel := context.WithTimeout(ctx, maxWaitInterval)
-		defer cancel()
+	query := fmt.Sprintf("tm.event = 'Tx' AND tx.hash = '%s'", tx.TxHash)
+	resultEventChan, err := p.client.Subscribe(newCtx, "tx-result-waiter", query)
+	if err != nil {
+		return &types.TxResult{
+			Error: err,
+			TxResult: &relayTypes.TxResponse{
+				TxHash: tx.TxHash,
+			},
+		}, err
+	}
+	defer p.client.Unsubscribe(newCtx, "tx-result-waiter", query)
 
-		query := fmt.Sprintf("tm.event = 'Tx' AND tx.hash = '%s'", txHash)
-		resultEventChan, err := p.client.Subscribe(newCtx, "tx-result-waiter", query)
-		if err != nil {
-			txRes <- &types.TxResultChan{
+	for {
+		select {
+		case <-ctx.Done():
+			return &types.TxResult{TxResult: &relayTypes.TxResponse{TxHash: tx.TxHash}}, ctx.Err()
+		case e := <-resultEventChan:
+			eventDataJSON, err := jsoniter.Marshal(e.Data)
+			if err != nil {
+				return &types.TxResult{TxResult: &relayTypes.TxResponse{TxHash: tx.TxHash}}, err
+			}
+
+			txRes := new(types.TxResultWaitResponse)
+			if err := jsoniter.Unmarshal(eventDataJSON, txRes); err != nil {
+				return &types.TxResult{TxResult: &relayTypes.TxResponse{TxHash: tx.TxHash}}, err
+			}
+
+			res := &types.TxResult{
 				TxResult: &relayTypes.TxResponse{
-					TxHash: txHash,
+					Height:    txRes.Height,
+					TxHash:    tx.TxHash,
+					Codespace: txRes.Result.Codespace,
+					Data:      string(txRes.Result.Data),
 				},
-				Error: err,
 			}
-			return
-		}
-		defer p.client.Unsubscribe(newCtx, "tx-result-waiter", query)
-
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case e := <-resultEventChan:
-				eventDataJSON, err := jsoniter.Marshal(e.Data)
-				if err != nil {
-					txRes <- &types.TxResultChan{
-						TxResult: &relayTypes.TxResponse{
-							TxHash: txHash,
-						}, Error: err,
-					}
-					return
-				}
-
-				txWaitRes := new(types.TxResultWaitResponse)
-				if err := jsoniter.Unmarshal(eventDataJSON, txWaitRes); err != nil {
-					txRes <- &types.TxResultChan{
-						TxResult: &relayTypes.TxResponse{
-							TxHash: txHash,
-						}, Error: err,
-					}
-					return
-				}
-				if uint32(txWaitRes.Result.Code) != types.CodeTypeOK {
-					txRes <- &types.TxResultChan{
-						Error: fmt.Errorf(txWaitRes.Result.Log),
-						TxResult: &relayTypes.TxResponse{
-							Height:    txWaitRes.Height,
-							TxHash:    txHash,
-							Codespace: txWaitRes.Result.Codespace,
-							Code:      relayTypes.ResponseCode(txWaitRes.Result.Code),
-							Data:      string(txWaitRes.Result.Data),
-						},
-					}
-					return
-				}
-
-				txRes <- &types.TxResultChan{
-					TxResult: &relayTypes.TxResponse{
-						Height:    txWaitRes.Height,
-						TxHash:    txHash,
-						Codespace: txWaitRes.Result.Codespace,
-						Code:      relayTypes.ResponseCode(txWaitRes.Result.Code),
-						Data:      string(txWaitRes.Result.Data),
-					},
-				}
-				return
+			if uint32(txRes.Result.Code) != types.CodeTypeOK {
+				return res, fmt.Errorf("transaction failed with error: %v", txRes.Result.Log)
 			}
+			res.TxResult.Code = relayTypes.Success
+			return res, nil
 		}
-	}(txResChan)
-	return txResChan
+	}
 }
 
 func (p *Provider) MessageReceived(ctx context.Context, key *relayTypes.MessageKey) (bool, error) {
-	queryMsg := &types.QueryReceiptMsg{
-		GetReceipt: &types.GetReceiptMsg{
-			SrcNetwork: key.Src,
-			ConnSn:     strconv.FormatUint(key.Sn, 10),
-		},
-	}
-	rawQueryMsg, err := jsoniter.Marshal(queryMsg)
-	if err != nil {
-		return false, err
-	}
+	switch key.EventType {
+	case events.EmitMessage:
+		queryMsg := &types.QueryReceiptMsg{
+			GetReceipt: &types.GetReceiptMsg{
+				SrcNetwork: key.Src,
+				ConnSn:     key.Sn.String(),
+			},
+		}
+		rawQueryMsg, err := jsoniter.Marshal(queryMsg)
+		if err != nil {
+			return false, err
+		}
 
-	res, err := p.client.QuerySmartContract(ctx, p.cfg.Contracts[relayTypes.ConnectionContract], rawQueryMsg)
-	if err != nil {
-		p.logger.Error("failed to check if message is received: ", zap.Error(err))
-		return false, err
-	}
+		res, err := p.client.QuerySmartContract(ctx, p.cfg.Contracts[relayTypes.ConnectionContract], rawQueryMsg)
+		if err != nil {
+			p.logger.Error("failed to check if message is received: ", zap.Error(err))
+			return false, err
+		}
 
-	receiptMsgRes := types.QueryReceiptMsgResponse{}
-	return receiptMsgRes.Status, jsoniter.Unmarshal(res.Data, &receiptMsgRes.Status)
+		receiptMsgRes := types.QueryReceiptMsgResponse{}
+		return receiptMsgRes.Status, jsoniter.Unmarshal(res.Data, &receiptMsgRes.Status)
+	case events.CallMessage:
+		return false, nil
+	case events.RollbackMessage:
+		return false, nil
+	default:
+		return true, fmt.Errorf("unknown event type")
+	}
 }
 
 func (p *Provider) QueryBalance(ctx context.Context, addr string) (*relayTypes.Coin, error) {
@@ -451,7 +435,7 @@ func (p *Provider) FinalityBlock(ctx context.Context) uint64 {
 
 func (p *Provider) RevertMessage(ctx context.Context, sn *big.Int) error {
 	msg := &relayTypes.Message{
-		Sn:        sn.Uint64(),
+		Sn:        sn,
 		EventType: events.RevertMessage,
 	}
 	_, err := p.call(ctx, msg)
@@ -459,7 +443,7 @@ func (p *Provider) RevertMessage(ctx context.Context, sn *big.Int) error {
 }
 
 // SetFee
-func (p *Provider) SetFee(ctx context.Context, networkdID string, msgFee, resFee uint64) error {
+func (p *Provider) SetFee(ctx context.Context, networkdID string, msgFee, resFee *big.Int) error {
 	msg := &relayTypes.Message{
 		Src:       networkdID,
 		Sn:        msgFee,
@@ -502,8 +486,8 @@ func (p *Provider) SetAdmin(ctx context.Context, address string) error {
 // ExecuteRollback
 func (p *Provider) ExecuteRollback(ctx context.Context, sn *big.Int) error {
 	msg := &relayTypes.Message{
-		Sn:        sn.Uint64(),
-		EventType: events.ExecuteRollback,
+		Sn:        sn,
+		EventType: events.RollbackMessage,
 	}
 	_, err := p.call(ctx, msg)
 	return err
@@ -534,8 +518,8 @@ func (p *Provider) getHeightStream(done <-chan bool, fromHeight, toHeight uint64
 			select {
 			case <-done:
 				return
-			case heightChan <- &types.HeightRange{Start: fromHeight, End: fromHeight + 2}:
-				fromHeight += 2
+			case heightChan <- &types.HeightRange{Start: fromHeight, End: fromHeight + p.cfg.BlockBatchSize}:
+				fromHeight += p.cfg.BlockBatchSize
 			}
 		}
 	}(fromHeight, toHeight, heightChan)
@@ -639,7 +623,7 @@ func (p *Provider) getMessagesFromTxList(resultTxList []*coreTypes.ResultTx) ([]
 				p.logger.Info("Detected eventlog",
 					zap.Uint64("height", msg.MessageHeight),
 					zap.String("target_network", msg.Dst),
-					zap.Uint64("sn", msg.Sn),
+					zap.Uint64("sn", msg.Sn.Uint64()),
 					zap.String("event_type", msg.EventType),
 				)
 			}
@@ -672,7 +656,7 @@ func (p *Provider) getRawContractMessage(message *relayTypes.Message) (wasmTypes
 	case events.SetFee:
 		setFee := types.NewExecSetFee(message.Src, message.Sn, message.ReqID)
 		return jsoniter.Marshal(setFee)
-	case events.ExecuteRollback:
+	case events.RollbackMessage:
 		executeRollback := types.NewExecExecuteRollback(message.Sn)
 		return jsoniter.Marshal(executeRollback)
 	default:
@@ -684,7 +668,7 @@ func (p *Provider) getNumOfPipelines(diff int) int {
 	if diff <= runtime.NumCPU() {
 		return diff
 	}
-	return runtime.NumCPU() / 2
+	return runtime.NumCPU()
 }
 
 func (p *Provider) runBlockQuery(ctx context.Context, blockInfoChan chan *relayTypes.BlockInfo, fromHeight, toHeight uint64) uint64 {
@@ -693,7 +677,7 @@ func (p *Provider) runBlockQuery(ctx context.Context, blockInfoChan chan *relayT
 
 	heightStream := p.getHeightStream(done, fromHeight, toHeight)
 
-	diff := int(toHeight-fromHeight) / 2
+	diff := int(toHeight-fromHeight/p.cfg.BlockBatchSize) + 1
 
 	numOfPipelines := p.getNumOfPipelines(diff)
 	wg := &sync.WaitGroup{}
@@ -783,7 +767,7 @@ func (p *Provider) SubscribeMessageEvents(ctx context.Context, blockInfoChan cha
 				p.logger.Info("Detected eventlog",
 					zap.Int64("height", res.Height),
 					zap.String("target_network", msg.Dst),
-					zap.Uint64("sn", msg.Sn),
+					zap.Uint64("sn", msg.Sn.Uint64()),
 					zap.String("event_type", msg.EventType),
 				)
 			}