+++ utils/multisig/multisig_test.go
@@ -365,3 +365,71 @@ func TestTransferRune(t *testing.T) {
 	a, _ := json.Marshal(artifact)
 	fmt.Printf("Artifact: %s\n", string(a))
 }
+
+func TestTransferBitcoinWithBridgeMessage(t *testing.T) {
+	chainParam := &chaincfg.SigNetParams
+
+	inputs := []*UTXO{
+		{
+			IsRelayersMultisig: false,
+			TxHash:        "ae9f43a77d861d5076ebdb1af0d76af033843b784766a1d07a78a68fe845c012",
+			OutputIdx:     1,
+			OutputAmount:  3808,
+		},
+	}
+
+	outputs := []*OutputTx{
+		{
+			ReceiverAddress: "tb1pv5j5j0dmq2c8d0vnehrlsgrwr9g95m849dl5v0tal8chfdgzqxfskv0w8u",
+			Amount:          1000,
+		},
+	}
+
+	userPrivKeys, userMultisigInfo := randomMultisigInfo(2, 2, chainParam, []int{0, 3}, 1, 1)
+	userMultisigWallet, _ := BuildMultisigWallet(userMultisigInfo)
+
+	changeReceiverAddress := "tb1pgzx880yfr7q8dgz8dqhw50sncu4f4hmw5cn3800354tuzcy9jx5shvv7su"
+	msgTx, _, txSigHashes, _ := CreateMultisigTx(inputs, outputs, 200, &MultisigWallet{}, userMultisigWallet, chainParam, changeReceiverAddress, 1)
+
+	// Add Bridge Message
+
+	message := []byte("Test Bridge Message")
+	scripts, _ := CreateBridgeMessageScripts(message, 5)
+	for i, script := range scripts {
+		fmt.Println("OP_RETURN ", i, " script ", script)
+		msgTx.AddTxOut(wire.NewTxOut(0, script))
+	}
+
+
+	tapSigParams := TapSigParams {
+		TxSigHashes: txSigHashes,
+		RelayersPKScript: []byte{},
+		RelayersTapLeaf: txscript.TapLeaf{},
+		UserPKScript: userMultisigWallet.PKScript,
+		UserTapLeaf: userMultisigWallet.TapLeaves[1],
+	}
+
+	totalSigs := [][][]byte{}
+
+	// USER SIGN TX
+	userSigs, _ := PartSignOnRawExternalTx(userPrivKeys[1], msgTx, inputs, tapSigParams, chainParam, true)
+	totalSigs = append(totalSigs, userSigs)
+
+	// COMBINE SIGNS
+	signedMsgTx, err := CombineMultisigSigs(msgTx, inputs, userMultisigWallet, 0, userMultisigWallet, 1, totalSigs)
+
+	var signedTx bytes.Buffer
+	signedMsgTx.Serialize(&signedTx)
+	hexSignedTx := hex.EncodeToString(signedTx.Bytes())
+	signedMsgTxID := signedMsgTx.TxHash().String()
+
+	fmt.Println("hexSignedTx: ", hexSignedTx)
+	fmt.Println("signedMsgTxID: ", signedMsgTxID)
+	fmt.Println("err sign: ", err)
+
+	// Decode bridge message
+	decodedMessage, err := readRelayMessage(signedMsgTx, false)
+	fmt.Println("err decode: ", err)
+	fmt.Println("origina message: ", message)
+	fmt.Println("decoded message: ", decodedMessage)
+}